(2.1.8) - Optimizations
I'm happy to share that we've fine-tuned the script for better performance. I would like the think of this as the first step in our previously mentioned goals. During idle times, the impact on client performance is now between 0.00ms and 0.01ms, while previously, we saw numbers around 0.03ms. While driving, if the vehicle is above the fuel shutoff threshold, we've managed to maintain around ~0.01ms to ~0.02ms.

I want to be transparent – when you hit the shutoff threshold (configured in the settings), the performance impact still registers at around 0.03ms to 0.04ms, which is still lower than previously seen at ~0.05ms. I've worked diligently to push this lower, but reaching my ideal target has proven challenging. Rest assured, I've done my best to minimize this impact, and I will continue to try and see ways to lower this even more.

Again, like always I want to make it clear that for any issues, big or small, you are always welcome to open up a ticket in Discord or create an issue on GitHub:
https://github.com/CodineDev/cdn-fuel/issues/new/choose